requirements.txt
bimmer-connected==0.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
twilio==7.3.1
urllib3==1.26.7
python-dotenv==0.20.0
